refactor: create utils file and data directory for theme colors
