+++ docs/src/components/ThemeColorsTable.tsx
@@ -7,40 +7,11 @@ import TabItem from '@theme/TabItem';
 //@ts-ignore
 import Tabs from '@theme/Tabs';
 
-type DataObject = {
-  [key: string]: string | DataObject;
-};
-
-const getUniqueNestedKeys = (data: DataObject): string[] => {
-  const keys: string[] = [];
-
-  const traverseObject = (obj: DataObject) => {
-    for (const [key, value] of Object.entries(obj)) {
-      if (typeof value === 'object') {
-        traverseObject(value);
-      } else {
-        keys.push(key);
-      }
-    }
-  };
-
-  traverseObject(data);
-
-  return [...new Set(keys)];
-};
-
-const getNestingLevel = (obj: DataObject): number => {
-  let maxNestedLevel = 0;
-
-  for (const key in obj) {
-    if (typeof obj[key] === 'object') {
-      const nestedLevel = getNestingLevel(obj[key] as DataObject) + 1;
-      maxNestedLevel = Math.max(maxNestedLevel, nestedLevel);
-    }
-  }
-
-  return maxNestedLevel;
-};
+import {
+  DataObject,
+  getMaxNestedLevel,
+  getUniqueNestedKeys,
+} from '../utils/themeColors';
 
 const getTableHeader = (keys: string[]): JSX.Element[] => {
   return keys.map((key) => <th key={key}>{key}</th>);
@@ -68,7 +39,7 @@ const ThemeColorsTable = ({
   }
 
   const uniqueKeys = getUniqueNestedKeys(data);
-  const nestingLevel = getNestingLevel(data);
+  const nestingLevel = getMaxNestedLevel(data);
 
   if (nestingLevel === 1) {
     const rows = Object.keys(data).map((mode) => {
